Improve page query performance
Somehow querying the original of of childImageSharp takes quite long.
Getting the height and width of the gatsbyImageData is much faster. The
absolute values are different, but as we are only interested in the
aspect ratio, these values can be used instead.

Before this change I got ~1.2 queries/s, now it is ~5 queries/s. Total
build time (with filled cache) dropped by a factor of 2.3.